refactor: move extension manifest to separate file
