+++ vicinae/src/services/extension-registry/extension-registry.cpp
@@ -1,10 +1,7 @@
 #include "extension-registry.hpp"
-#include "common.hpp"
 #include "services/local-storage/local-storage-service.hpp"
-#include "vicinae.hpp"
 #include <QJsonArray>
 #include "services/extension-registry/extension-registry.hpp"
-#include "utils/utils.hpp"
 #include "xdgpp/env/env.hpp"
 #include "zip/unzip.hpp"
 #include <QtConcurrent/qtconcurrentrun.h>
@@ -22,12 +19,11 @@ namespace fs = std::filesystem;
 ExtensionRegistry::ExtensionRegistry(LocalStorageService &storage) : m_storage(storage) {
   using namespace std::chrono_literals;
 
-  m_paths = xdgpp::dataDirs() | std::views::transform([](auto &&p) { return p / "vicinae" / "extensions"; }) |
-            std::ranges::to<std::vector>();
+  m_extDirs = extensionDirectories();
   m_rescanDebounce.setInterval(100ms);
   m_rescanDebounce.setSingleShot(true);
 
-  for (const auto &dir : extensionDirs()) {
+  for (const auto &dir : m_extDirs) {
     m_watcher->addPath(dir.c_str());
   }
 
@@ -38,39 +34,9 @@ ExtensionRegistry::ExtensionRegistry(LocalStorageService &storage) : m_storage(s
   connect(&m_rescanDebounce, &QTimer::timeout, this, [this]() { requestScan(); });
 }
 
-CommandArgument ExtensionRegistry::parseArgumentFromObject(const QJsonObject &obj) {
-  CommandArgument arg;
-  QString type = obj.value("type").toString();
-
-  if (type == "text") arg.type = CommandArgument::Text;
-  if (type == "password") arg.type = CommandArgument::Password;
-  if (type == "dropdown") arg.type = CommandArgument::Dropdown;
-
-  arg.name = obj.value("name").toString();
-  arg.placeholder = obj.value("placeholder").toString();
-  arg.required = obj.value("required").toBool(false);
-
-  if (type == "dropdown") {
-    auto data = obj.value("data").toArray();
-    std::vector<CommandArgument::DropdownData> options;
-
-    options.reserve(data.size());
-
-    for (const auto &child : data) {
-      auto obj = child.toObject();
-
-      options.push_back({.title = obj["title"].toString(), .value = obj["value"].toString()});
-    }
-
-    arg.data = options;
-  }
-
-  return arg;
-}
-
 QFuture<bool> ExtensionRegistry::installFromZip(const QString &id, std::string data,
                                                 std::function<void(bool)> cb) {
-  fs::path extractDir = localExtensionDir() / id.toStdString();
+  fs::path extractDir = localExtensionDirectory() / id.toStdString();
   auto future = QtConcurrent::run([id, data, extractDir]() {
     Unzipper unzip = std::string_view(data);
 
@@ -127,87 +93,16 @@ bool ExtensionRegistry::uninstall(const QString &id) {
   return true;
 }
 
-Preference ExtensionRegistry::parsePreferenceFromObject(const QJsonObject &obj) {
-  auto type = obj["type"].toString();
-  Preference base;
-
-  base.setTitle(obj["title"].toString());
-  base.setDescription(obj["description"].toString());
-  base.setName(obj["name"].toString());
-  base.setPlaceholder(obj["placeholder"].toString());
-  base.setRequired(obj["required"].toBool());
-  base.setDefaultValue(obj.value("default"));
-
-  if (type == "textfield") {
-    base.setData(Preference::TextData());
-  } else if (type == "password") {
-    base.setData(Preference::PasswordData());
-  } else if (type == "checkbox") {
-    auto checkbox = Preference::CheckboxData(obj["label"].toString());
-    base.setData(checkbox);
-  } else if (type == "appPicker") {
-    base.setData(Preference::AppPickerData());
-  } else if (type == "file") {
-    base.setData(Preference::FilePickerData());
-  } else if (type == "directory") {
-    base.setData(Preference::DirectoryPickerData());
-  } else if (type == "dropdown") {
-    auto data = obj["data"].toArray();
-    std::vector<Preference::DropdownData::Option> options;
-
-    options.reserve(data.size());
-
-    for (const auto &child : data) {
-      auto obj = child.toObject();
-
-      options.push_back({.title = obj["title"].toString(), .value = obj["value"].toString()});
-    }
-
-    base.setData(Preference::DropdownData{options});
-  } else {
-    qWarning() << "Unknown extension preference type" << type;
-  }
-
-  return base;
-}
-
-ExtensionManifest::Command ExtensionRegistry::parseCommandFromObject(const QJsonObject &obj) {
-  ExtensionManifest::Command command;
-  auto type = obj.value("mode");
-
-  command.name = obj.value("name").toString();
-  command.title = obj.value("title").toString();
-  command.description = obj.value("description").toString();
-  command.defaultDisabled = obj.value("disabledByDefault").toBool(false);
-
-  if (obj.contains("icon")) { command.icon = obj.value("icon").toString(); }
-
-  if (type == "view") {
-    command.mode = CommandMode::CommandModeView;
-  } else if (type == "no-view") {
-    command.mode = CommandMode::CommandModeNoView;
-  } else {
-    command.mode = CommandMode::CommandModeInvalid;
-  }
-
-  for (const auto &obj : obj.value("preferences").toArray()) {
-    command.preferences.emplace_back(parsePreferenceFromObject(obj.toObject()));
-  }
-
-  for (const auto &obj : obj.value("arguments").toArray()) {
-    command.arguments.emplace_back(parseArgumentFromObject(obj.toObject()));
-  }
-
-  return command;
-}
-
-fs::path ExtensionRegistry::localExtensionDir() { return xdgpp::dataHome() / "vicinae" / "extensions"; }
+fs::path ExtensionRegistry::localExtensionDirectory() { return xdgpp::dataHome() / "vicinae" / "extensions"; }
 
 fs::path ExtensionRegistry::supportDirectory() { return xdgpp::dataHome() / "vicinae" / "support"; }
 
 fs::path ExtensionRegistry::supportDirectory(const std::string &id) { return supportDirectory() / id; }
 
-std::span<const fs::path> ExtensionRegistry::extensionDirs() const { return m_paths; }
+std::vector<fs::path> ExtensionRegistry::extensionDirectories() {
+  return xdgpp::dataDirs() | std::views::transform([](auto &&p) { return p / "vicinae" / "extensions"; }) |
+         std::ranges::to<std::vector>();
+}
 
 std::vector<ExtensionManifest> ExtensionRegistry::scanAll() {
   std::error_code ec;
@@ -216,7 +111,7 @@ std::vector<ExtensionManifest> ExtensionRegistry::scanAll() {
   manifests.reserve(m_installed.size());
   m_installed.clear();
 
-  for (const fs::path &path : extensionDirs()) {
+  for (const fs::path &path : m_extDirs) {
     for (const auto &entry : fs::directory_iterator(path, ec)) {
       if (!entry.is_directory(ec)) continue;
 
@@ -232,7 +127,7 @@ std::vector<ExtensionManifest> ExtensionRegistry::scanAll() {
         continue;
       }
 
-      auto manifest = scanBundle(entry.path());
+      auto manifest = ExtensionManifest::fromPackageJson(entry.path());
 
       if (!manifest) {
         qCritical() << "Failed to load bundle at" << entry.path().c_str() << manifest.error().m_message;
@@ -249,68 +144,4 @@ std::vector<ExtensionManifest> ExtensionRegistry::scanAll() {
 
 bool ExtensionRegistry::isInstalled(const QString &id) const {
   return m_installed.contains(id.toStdString());
-}
-
-tl::expected<ExtensionManifest, ManifestError> ExtensionRegistry::scanBundle(const fs::path &path) {
-  static const std::set<CommandMode> supportedModes{CommandMode::CommandModeView, CommandModeNoView};
-  fs::path manifestPath = path / "package.json";
-
-  if (!fs::exists(manifestPath)) {
-    return tl::unexpected<ManifestError>(
-        QString("Could not find package.json file at %1").arg(manifestPath.c_str()));
-  }
-
-  QFile file(manifestPath);
-
-  if (!file.open(QIODevice::ReadOnly)) {
-    return tl::unexpected<ManifestError>(QString("Failed to open %1 for read").arg(manifestPath.c_str()));
-  }
-
-  QJsonParseError jsonError;
-  auto json = QJsonDocument::fromJson(file.readAll(), &jsonError);
-
-  if (jsonError.error) {
-    return tl::unexpected<ManifestError>(
-        QString("Failed to parse package.json at %1").arg(manifestPath.c_str()));
-  }
-
-  ExtensionManifest manifest;
-  auto obj = json.object();
-  auto deps = obj.value("dependencies").toObject();
-
-  manifest.path = path;
-  manifest.id = QString::fromStdString(getLastPathComponent(path));
-  manifest.name = obj.value("name").toString();
-  manifest.title = obj.value("title").toString();
-  manifest.description = obj.value("description").toString();
-  manifest.icon = obj.value("icon").toString();
-  manifest.author = obj.value("author").toString();
-  manifest.needsRaycastApi = deps.contains(Omnicast::RAYCAST_NPM_API_PACKAGE);
-
-  if (manifest.id.startsWith("store.vicinae.")) {
-    manifest.provenance = ExtensionManifest::Provenance::Vicinae;
-  } else if (manifest.id.startsWith("store.raycast.")) {
-    manifest.provenance = ExtensionManifest::Provenance::Raycast;
-  } else {
-    manifest.provenance = ExtensionManifest::Provenance::Local;
-  }
-
-  for (const auto &obj : obj.value("categories").toArray()) {
-    manifest.categories.emplace_back(obj.toString());
-  }
-
-  for (const auto &obj : obj.value("commands").toArray()) {
-    auto command = parseCommandFromObject(obj.toObject());
-
-    command.provenance = manifest.provenance;
-    command.entrypoint = path / std::format("{}.js", command.name.toStdString());
-
-    if (supportedModes.contains(command.mode)) { manifest.commands.emplace_back(command); }
-  }
-
-  for (const auto &obj : obj.value("preferences").toArray()) {
-    manifest.preferences.emplace_back(parsePreferenceFromObject(obj.toObject()));
-  }
-
-  return manifest;
 }
